+++ module-system/module-implementations/sov-sequencer-registry/src/tests/mod.rs
@@ -1,0 +1,5 @@
+pub mod helpers;
+
+mod hooks;
+mod registration_mechanism;
+mod slashing_mechanism;