+++ CHANGELOG.md
@@ -29,4 +29,5 @@ This breaks the `StateTransitionFunction` API but the breaking changes don't pro
 
 - #443 Removes the `coins` field in the `SequencerRegistry` struct. It is replaced by a `minimum_bond` field and the `TokenId` becomes `GAS_TOKEN_ID`. The configuration structure `SequencerRegistryConfig` should be updated to replace the `coin` field by the new `minimum_bond` field.
 
+- #444 Moves the tests for the `SequencerRegistry` module to the `src` directory of the same crate.